remove all the outdated/unused code

+++ src/main/java/meteordevelopment/meteorbot/command/commands/DonatorCommand.java
@@ -1,23 +1,26 @@
-package meteordevelopment.meteorbot.commands.moderator;
+package meteordevelopment.meteorbot.command.commands;
 
 import kong.unirest.Unirest;
 import meteordevelopment.meteorbot.Config;
 import meteordevelopment.meteorbot.MeteorBot;
-import meteordevelopment.meteorbot.commands.Category;
-import meteordevelopment.meteorbot.commands.Command;
-import meteordevelopment.meteorbot.utils.Utils;
+import meteordevelopment.meteorbot.command.Command;
+import net.dv8tion.jda.api.Permission;
 import net.dv8tion.jda.api.entities.Member;
 import net.dv8tion.jda.api.events.message.MessageReceivedEvent;
 
 public class DonatorCommand extends Command {
     public DonatorCommand() {
-        super(Category.Moderator, "Pinged person gets donator.", "donator");
+        super("donator");
     }
 
     @Override
     public void run(MessageReceivedEvent event) {
+        if (!event.getMember().hasPermission(Permission.MANAGE_ROLES)) {
+            event.getChannel().sendMessage("Moderator only command!").queue();
+            return;
+        }
+
         event.getMessage().delete().queue();
-        if (!Utils.onlyMod(event)) return;
 
         String[] split = event.getMessage().getContentRaw().split(" ");
         String id = null;
@@ -39,7 +42,7 @@ public class DonatorCommand extends Command {
             event.getChannel().sendMessage(member.getAsMention() + " thanks for donating to Meteor Client. You can go to https://meteorclient.com and create an account, link your Discord and Minecraft accounts to get a cape. You can also upload a custom one. Also be sure send your Minecraft name so we can give you donator role on our pvp server. (pvp.meteorclient.com)").queue();
 
             Unirest.post("https://meteorclient.com/api/discord/giveDonator")
-                .header("Authorization", Config.TOKEN)
+                .header("Authorization", Config.BACKEND_TOKEN)
                 .queryString("id", member.getId())
                 .asEmpty();
         });
